Add native macOS ARM64 (Apple Silicon) build support for DeepVariant
This is a pioneering effort - DeepVariant has never been built natively on
macOS before (only Ubuntu 22.04 is officially supported).

Key changes:
- Create build-prereq-macos.sh and run-prereq-macos.sh for Homebrew-based
  dependency management (replacing apt-get)
- Add macOS/ARM64 detection in settings.sh with proper compiler flags
- Fix htslib.BUILD for macOS (no fdatasync, no SSE, .dylib plugin ext)
- Add .bazelrc macOS config (macos_arm64, Boost v1 process API)
- Fix build_release_binaries.sh for BSD tools (sed -i '', Python relpath
  instead of GNU ln --relative)
- Add Boost as Bazel repository (third_party/boost.BUILD + WORKSPACE)
- Fix Boost 1.90 v1/v2 process API split in fast_pipeline.{h,cc}
- Add @boost//:boost deps to all targets needing Boost headers
- Fix int64_t/long type mismatches for ARM64 (0L -> static_cast<int64_t>(0))
- Fix libssw.BUILD to include sse2neon.h for ARM64 SSE translation
- Fix sdsl_lite.BUILD include path conflict with C++20 <version> header
- Add CLIF C++ runtime abseil deps to clif.BUILD
- Update Python runtime to use Homebrew python3.10

Tested: all binaries (make_examples, call_variants, postprocess_variants,
train, show_examples, etc.) build and run successfully on M1 Max.

+++ deepvariant/allelecounter.cc
@@ -307,7 +307,7 @@ void AlleleCounter::Init() {
   auto full_interval_offset = interval_.start() - reads_interval_.start();
   // If interval_ starts before reads_interval_ start then we don't need to
   // offset reference bases.
-  full_interval_offset = std::max(full_interval_offset, 0L);
+  full_interval_offset = std::max(full_interval_offset, static_cast<int64_t>(0));
   for (int i = 0; i < len; ++i) {
     AlleleCount allele_count;
     const int64_t pos = interval_.start() + i;
